requirements.txt
# Core Deep Learning
torch>=2.0.0
transformers>=4.30.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# HTML Cleaning
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Progress Bars
tqdm>=4.65.0

# Hugging Face
huggingface_hub>=0.16.0
datasets>=2.14.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Optional (for visualization)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Optional (for serving)
fastapi>=0.100.0
uvicorn>=0.23.0